release 5.4.1.4

+++ src/service/src/XCart/Bus/System/ResourceChecker.php
@@ -86,7 +86,7 @@ class ResourceChecker
      */
     private function getTime(): int
     {
-        return (microtime(true) * 10000);
+        return (int)(microtime(true) * 10000);
     }
 
     /**
